feat: make filendir methods behave like equivalent fs.write methods
The goal of this change is to, in addition to create missing directories,
provide the exact same features and behaviour as node
fs.writeFile and fs.writeFileSync.

I believe it is a more correct approach.

This includes:

* methods have the same signatures than corresponding fs.writeFile ones.
* method writeFile has been added as synonym for async. write
* docs have been updated
* more unit tests

BREAKING CHANGE:

* synchronous write errors are not handled anymore.

They will simply (re)throw an error.
In opposition to the previous behaviour which would return
true or false depending on success. You now have to handle the errors
as you would for the fs.writeFileSync method.